requirements.txt
requests==2.28.2
pandas==1.5.0
openpyxl==3.0.10
Pillow==9.5.0
numpy==1.21.0